Minimize the set of recomputed errors after each edge collapse.
Don't add edges that couldn't have been affected by an edge collapse to the set of dirty edges.
There are still some redundant computations done, but even removing 1 iteration from the inner
loop reduces CLOD build time for larger meshes by around 20% (single threaded).

Fix missing face geometry copy during mesh compaction causing incorrect geometry indices in the
output meshlets.

+++ example/main.cpp
@@ -353,7 +353,7 @@ static void ParallelForCallback(void* /*user_data*/, void* mdt_data, u32 work_it
 	struct Iterator {
 		int64_t index = 0;
 		
-		using value_type = double;
+		using value_type = int64_t;
 		using difference_type = int64_t;
 		using pointer = int64_t*;
 		using reference = int64_t&;
